+++ Fiddlehead/Settings/KeychainManager.swift
@@ -4,7 +4,6 @@ import Security
 enum KeychainKey: String {
     case deepgramAPIKey = "deepgram_api_key"
     case claudeAPIKey = "claude_api_key"
-    case licenseKey = "license_key"
 }
 
 enum KeychainError: Error {